requirements-dev.txt
pytest==3.7.1
pytest-asyncio==0.10.0
pytest-django==3.3.3
pytest-cov==2.5.1
responses==0.9.0
flake8==3.5.0
black
pre-commit
freezegun
isort==4.3.4